requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
bcrypt==4.2.1
python-dotenv==1.0.1
PyJWT==2.10.1
Werkzeug==3.1.3
boto3>=1.34.0
